Keep original project structure (#2)
Original projects are not modified, only filtered.
The toctrees are expected to support the given file split - that's the project's task.
The needimports for `needs_json_labels` are generated to
```
needimports/docs/index.rst
needimports/docs/<project>_needs.rst
```
so the root index.rst must include a glob for this:
```
*/docs/index
```

The project structure must follow this pattern as well, at least on root level.

The `config_setting` are removed as all is based on targets and file groups now.
The user is in charge to properly layout their projects and targets.
Sphinx will complain if existing files are not included or missing files are mentioned in toctrees.